@@ -1,6 +1,6 @@
 {
 	"name": "@notwoods/logo",
-	"version": "1.1.1",
+	"version": "2.0.0",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build && npm run package",
@@ -29,29 +29,29 @@
 		"svg/*.svg"
 	],
 	"peerDependencies": {
-		"svelte": "^4.0.0 || ^5.0.0-next.0"
+		"svelte": "^5.0.0"
 	},
 	"devDependencies": {
-		"@storybook/addon-essentials": "^8.3.4",
-		"@storybook/addon-interactions": "^8.3.4",
-		"@storybook/addon-svelte-csf": "^5.0.0-next.3",
-		"@storybook/svelte": "^8.3.4",
-		"@storybook/sveltekit": "^8.3.4",
-		"@storybook/test": "^8.3.4",
-		"@sveltejs/adapter-static": "^3.0.5",
-		"@sveltejs/kit": "^2.5.27",
-		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.0",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.2.6",
-		"publint": "^0.1.9",
-		"storybook": "^8.0.9",
-		"svelte": "^5.0.0-next.0",
-		"svelte-check": "^4.0.0",
-		"svgo": "^3.2.0",
-		"tslib": "^2.4.1",
-		"typescript": "^5.5.0",
-		"vite": "^5.4.4"
+		"@storybook/addon-essentials": "^8.4.1",
+		"@storybook/addon-interactions": "^8.4.1",
+		"@storybook/addon-svelte-csf": "^5.0.0-next.10",
+		"@storybook/svelte": "^8.4.1",
+		"@storybook/sveltekit": "^8.4.1",
+		"@storybook/test": "^8.4.1",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/kit": "^2.7.4",
+		"@sveltejs/package": "^2.3.7",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
+		"publint": "^0.2.12",
+		"storybook": "^8.4.1",
+		"svelte": "^5.1.9",
+		"svelte-check": "^4.0.5",
+		"svgo": "^3.3.2",
+		"tslib": "^2.8.1",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.10"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
